updated index.spec.ts file & other updates
