+++ examples/generate-all-models-within-the-same-file/.gitignore
